requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
openai
requests>=2.26.0